+++ cmd/firenear/cli/receipt-index-builder.go
@@ -23,8 +23,6 @@ func init() {
 		Description: "Produces a receipt index for a given set of blocks",
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().Uint64("receipt-index-builder-index-size", 10000, "size of receipt index bundles that will be created")
-			cmd.Flags().IntSlice("receipt-index-builder-lookup-index-sizes", []int{1000000, 100000, 10000, 1000}, "index bundle sizes that we will look for on start to find first unindexed block")
-			cmd.Flags().String("receipt-index-builder-index-store-url", "", "url of the index store")
 			cmd.Flags().Uint64("receipt-index-builder-start-block", 0, "block number to start indexing")
 			cmd.Flags().Uint64("receipt-index-builder-stop-block", 0, "block number to stop indexing")
 			return nil
@@ -35,7 +33,7 @@ func init() {
 		FactoryFunc: func(runtime *launcher.Runtime) (launcher.App, error) {
 			sfDataDir := runtime.AbsDataDir
 
-			indexStoreURL := MustReplaceDataDir(sfDataDir, viper.GetString("receipt-index-builder-index-store-url"))
+			indexStoreURL := MustReplaceDataDir(sfDataDir, viper.GetString("common-index-store-url"))
 			blockStoreURL := MustReplaceDataDir(sfDataDir, viper.GetString("common-merged-blocks-store-url"))
 
 			indexStore, err := dstore.NewStore(indexStoreURL, "", "", false)
@@ -44,7 +42,7 @@ func init() {
 			}
 
 			var lookupIdxSizes []uint64
-			lookupIndexSizes := viper.GetIntSlice("receipt-index-builder-lookup-index-sizes")
+			lookupIndexSizes := viper.GetIntSlice("common-block-index-sizes")
 			for _, size := range lookupIndexSizes {
 				if size < 0 {
 					return nil, fmt.Errorf("invalid negative size for bundle-sizes: %d", size)
@@ -72,8 +70,8 @@ func init() {
 			stopBlockNum := viper.GetUint64("receipt-index-builder-stop-block")
 
 			receiptIndexer := transform.NewNearBlockIndexer(indexStore, viper.GetUint64("receipt-index-builder-index-size"))
-			handler := bstream.HandlerFunc(func(blk *bstream.Block, obj interface{}) error {
-				receiptIndexer.ProcessBlock(blk.ToNative().(*pbnear.Block))
+			handler := bstream.HandlerFunc(func(blk *bstream.Block, _ interface{}) error {
+				receiptIndexer.ProcessBlock(blk.ToProtocol().(*pbnear.Block))
 				return nil
 			})